refactor: bottom sheet component
