+++ app/src/main/java/com/skgtecnologia/sisem/domain/login/model/LegalModel.kt
@@ -1,10 +1,10 @@
-package com.valkiria.uicomponents.components.bottomsheet
+package com.skgtecnologia.sisem.domain.login.model
 
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.painter.Painter
 import com.valkiria.uicomponents.props.TextStyle
 
-data class BottomSheetUiModel(
+data class LegalModel(
     val icon: Painter? = null,
     val title: String,
     val titleTextStyle: TextStyle,
