+++ app/src/main/java/com/skgtecnologia/sisem/ui/login/LoginScreen.kt
@@ -13,7 +13,7 @@ import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.constraintlayout.compose.Dimension
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.skgtecnologia.sisem.domain.login.model.LoginLink
-import com.skgtecnologia.sisem.domain.login.model.toBottomSheetUiModel
+import com.skgtecnologia.sisem.domain.login.model.toLegalModel
 import com.skgtecnologia.sisem.ui.navigation.AuthNavigationRoute
 import com.skgtecnologia.sisem.ui.sections.BodySection
 import com.valkiria.uicomponents.action.LoginUiAction
@@ -35,7 +35,7 @@ fun LoginScreen(
     val uiState = viewModel.uiState
 
     val sheetState = rememberModalBottomSheetState(
-        skipPartiallyExpanded = false
+        skipPartiallyExpanded = true
     )
     val scope = rememberCoroutineScope()
 
@@ -83,7 +83,9 @@ fun LoginScreen(
         }
 
         BottomSheetComponent(
-            uiModel = link.toBottomSheetUiModel(),
+            content = {
+                LegalContent(uiModel = link.toLegalModel())
+            },
             sheetState = sheetState,
             scope = scope
         ) {
