+++ uicomponents/src/main/java/com/valkiria/uicomponents/components/bottomsheet/BottomSheetComponent.kt
@@ -1,15 +1,8 @@
 package com.valkiria.uicomponents.components.bottomsheet
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.SheetState
@@ -21,19 +14,14 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.valkiria.uicomponents.mocks.getLoginTermsBottomSheetUiModel
-import com.valkiria.uicomponents.props.toTextStyle
 import com.valkiria.uicomponents.theme.UiComponentsTheme
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import timber.log.Timber
-import androidx.compose.ui.text.TextStyle.Companion as ComposeTextStyle
 
 @Composable
 fun BottomSheetComponent(
-    uiModel: BottomSheetUiModel,
-    modifier: Modifier = Modifier,
+    content: @Composable () -> Unit,
     sheetState: SheetState = rememberModalBottomSheetState(),
     scope: CoroutineScope = rememberCoroutineScope(),
     onAction: () -> Unit
@@ -50,43 +38,7 @@ fun BottomSheetComponent(
             containerColor = Color.DarkGray,
             scrimColor = MaterialTheme.colorScheme.background.copy(alpha = 0.8f)
         ) {
-            Row(
-                modifier = Modifier.padding(20.dp),
-                horizontalArrangement = Arrangement.SpaceAround
-            ) {
-                uiModel.icon?.let {
-                    Icon(
-                        painter = it,
-                        contentDescription = null,
-                        modifier = Modifier
-                            .padding(end = 16.dp)
-                            .size(42.dp),
-                        tint = MaterialTheme.colorScheme.primary
-                    )
-                }
-                Text(
-                    text = uiModel.title,
-                    style = uiModel.titleTextStyle.toTextStyle()
-                )
-            }
-            uiModel.subtitle?.let {
-                Text(
-                    text = it,
-                    modifier = Modifier.padding(20.dp),
-                    style = uiModel.subtitleTextStyle?.toTextStyle() ?: ComposeTextStyle.Default
-                )
-            }
-            Column(
-                modifier = Modifier
-                    .verticalScroll(rememberScrollState())
-                    .fillMaxSize()
-            ) {
-                Text(
-                    text = uiModel.text,
-                    modifier = Modifier.padding(20.dp),
-                    style = uiModel.textStyle.toTextStyle()
-                )
-            }
+            content()
         }
     }
 }
@@ -112,7 +64,13 @@ fun BottomSheetComponentPreview() {
             }
 
             BottomSheetComponent(
-                uiModel = getLoginTermsBottomSheetUiModel(),
+                content = {
+                    Text(
+                        text = """1. INTRODUCCIÓN\n\nLa Política de Seguridad de la Información del 
+            |sitio web y términos de uso del Sitio Web de la Secretaría Distrital de 
+            |Salud y el Fondo Financiero ...""".trimMargin()
+                    )
+                },
                 sheetState = sheetState,
                 scope = scope
             ) {
